// Public/src/models/HotSelling.Model.ts
export interface IHotSellingModel {
    Active: boolean
    BrandMasterId: number
    BrandMasterName: string
    Created: any
    CreatedBy: number
    Id: number
    IsValid: boolean
    MaximumValue: number
    MinimumValue: number
    Modified: any
    ModifiedBy: number
    Name: string
    OrderWishlistId: any
    ProductTypeId: number
    ProductTypeName: string
    RowOrder: number
    SeriesModelId: number
    SeriesModelName: string
    ThumbnailPath: string
    ValidationErrors: {}
}